feat: Add validation for booking start date and time
This commit adds validation to the create_booking endpoint in the router.py file. It checks if the start date and time provided for a booking is at least 72 hours in the future. If the start date and time is less than 72 hours from the current date and time, a 400 Bad Request response is returned with the message "Cannot book a lesson before 72 hours from the start date". This validation ensures that bookings can only be made with a minimum notice of 72 hours.