refactor: Update user fields in v01_schemas.py and router.py
The commit updates the user fields in the v01_schemas.py and router.py files. It adds the "name" and "surname" fields to the User and UpdateUserSchema classes in v01_schemas.py, as well as the UpdateStudentSchema and UpdateTeacherSchema classes. In router.py, it updates the update_student and update_teacher_school_subject functions to handle the new fields and update the corresponding user attributes. This change allows for more comprehensive user information management.