refactor: Add user_otp table for storing user OTP information
This commit adds a new table called "user_otp" to the database schema. The table has columns for the user ID, OTP (one-time password), and expiration datetime. This table will be used to store OTP information for password reset functionality. The commit also includes the necessary model definition and API endpoints for generating and validating OTPs.